Added EncodeString() and the Cipher type.
Added the EncodeString as a quick way to encode a string using
default params, just like other encoding packages in Go offer.

Also added a new Cipher type, which is useful when storing
ECE-encoded ciphers in a database. More specifically, this new
type offers basic validation when parsing JSON or raw database
values. Operations fail if the []byte value does not start
with a valid header.